+++ examples/dpfp_threaded.py
@@ -24,9 +24,14 @@
 # License along with this library; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-DPFP_THREADED = True
+import sys
+import importlib
 
-from dpfp import *
+import dpfp
+dpfp.DPFP_THREADED = True
+importlib.reload(dpfp)
+del dpfp
+from dpfp import main
 
 
 if __name__.rpartition(".")[-1] == "__main__":